backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic[email]>=2.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
